Renames src folder and moves test folder out of src.
